+++ tests/TestCase.php
@@ -30,6 +30,7 @@ abstract class TestCase extends OrchestraTestCase
         return [
             StatamicServiceProvider::class,
             \Goldnead\BrandContext\ServiceProvider::class,
+            \Goldnead\Suppression\ServiceProvider::class,
             \Goldnead\Leadhub\ServiceProvider::class,
             \Goldnead\Marketing\ServiceProvider::class,
         ];